requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord==1.7.3
discord.py==2.0.1
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.2
numpy==1.23.4
pandas==1.5.1
pbr==5.11.0
platformdirs==2.5.4
#psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stevedore==4.1.1
tzdata==2022.7
urllib3==1.26.12
use==0.1.0
virtualenv==20.16.7
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
yarl==1.8.1
